Fix the bug which prevents suggesting words
The code intended to show the user 'Potential Matches' when the entry
is not found on Webster's dictionary. The feature is valuable since it
guides the user to the nearest valid words. Still, there is a
mistake in the condition that prevents it to be met at the right
moments. Fix the condition to let the add-on notify users of the
suggested words.